Upload App/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
assemblyai==0.33.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
distro==1.9.0
ffmpeg==1.4
filelock==3.16.1
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
openai==1.47.1
platformdirs==4.3.6
pocketsphinx==5.0.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
sniffio==1.3.1
sounddevice==0.5.0
SpeechRecognition==3.10.4
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.6
websockets==13.1
Werkzeug==3.0.4